+++ CHANGELOG.md
@@ -6,10 +6,14 @@
 
 ### current: 0.2.0 (2022 Aug xx)
 
+ - new: `TimeWrapper` interface for time unit classes
  - new: `Timestamp` wrapper around timestamps
  - update: `Timespan`: get durations with specified units only
  - update: `Timespan`: add month unit
  - update: `Timespan`: **symbolType** renamed to **symbolFormat**
+ - update: `Timespan`: moved moment functions to timestamp
+ - update: `Timespan`: handles historic/negative values
+ - update: `Timespan`: format now returns duration
 
 ### 0.1.0 (2022 Aug 13)