+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 ## History
 
-### current: 0.2.0 (2022 Aug xx)
+### 0.2.0 (2022 Sep 22)
 
  - new: `TimeWrapper` interface for time unit classes
  - new: `Timestamp` wrapper around timestamps
+ - new: `abs` absolute value function added to both classes
  - update: `Timespan`: get durations with specified units only
  - update: `Timespan`: add month unit
  - update: `Timespan`: **symbolType** renamed to **symbolFormat**